find_unsafe2: refactor: move crate processing logic into lib.rs
